broke the project into modules and made changes to commits and methods
Rewrote some commits to English and fixed some logical errors. I also split the project into modules such as task.js and taskTracker.js and index.js.
